requirements.txt
requests
pyserial
gps3
